perf(worldgen): shrink ImprovedNoise permutation table to [256]uint8 (8x cache footprint)
ImprovedNoise.p was [256]int (2 KiB) holding a shuffle of 0..255. Chunk generation
interleaves dozens of ImprovedNoise instances (octave stacks) that compete for cache,
and sampleAndLerp does 14 pp() lookups per sample, so the oversized permutation table
was a hot-path cache pressure source (per a live generation CPU profile). Change it to
[256]uint8 (256 B) — an 8x smaller working set for the hottest data structure.

BYTE-IDENTICAL (load-bearing): Java's field is byte[] (SIGNED); the bytecode p(int)
does baload (sign-extends, e.g. stored 200 -> -56) then `& 255` (restores 200). The Go
uint8 store holds the value unsigned, so int(p[i&255]) yields 0..255 directly and the
`& 255` in pp() is a no-op producing the identical unsigned index. Values are always in
[0,255] so uint8(i) never truncates; the Fisher-Yates swap moves the same values. Every
permutation index, gradient hash, and output float is bit-for-bit unchanged.

CITE: net.minecraft.world.level.levelgen.synth.ImprovedNoise (p field / p(int) baload+iand).
A flat-gradient experiment was measured perf-neutral and NOT shipped; only the uint8 change.
Micro-bench (48 interleaved noises, round-robin): 307ns -> 213ns/op (~30% faster).
Gate: world suite green (byte-identical generated-terrain lock passes).

+++ world/levelgen/synth/improved.go
@@ -66,8 +66,16 @@ func gradDot(hash int, x, y, z float64) float64 {
 // ImprovedNoise is a single-octave improved-Perlin noise.
 // Source: net.minecraft.world.level.levelgen.synth.ImprovedNoise.
 type ImprovedNoise struct {
-	p          [256]int // the permutation table
-	xo, yo, zo float64  // the per-instance offsets
+	// OPTIMIZATION (byte-identical): the permutation table stores only values in
+	// [0,255] (it is a shuffle of 0..255), so it is held as [256]uint8 (256 bytes)
+	// instead of [256]int (2 KiB). sampleAndLerp issues 14 pp() lookups per sample
+	// and the generator interleaves dozens of ImprovedNoise instances, so the denser
+	// table shrinks the working set of the hottest data structure in noise generation
+	// 8x. pp() widens the stored byte back to int and re-applies the same `& 255`
+	// mask, so every returned permutation index — hence every gradient hash and every
+	// output float — is identical to the [256]int form.
+	p          [256]uint8 // the permutation table (values 0..255)
+	xo, yo, zo float64    // the per-instance offsets
 }
 
 // NewImprovedNoise seeds the permutation table + offsets from a RandomSource.
@@ -79,7 +87,7 @@ func NewImprovedNoise(r levelgen.RandomSource) *ImprovedNoise {
 	im.yo = r.NextDouble() * 256
 	im.zo = r.NextDouble() * 256
 	for i := 0; i < 256; i++ {
-		im.p[i] = i
+		im.p[i] = uint8(i) // i is 0..255, exact in uint8
 	}
 	for i := 0; i < 256; i++ {
 		j := int(r.NextIntN(int32(256 - i)))
@@ -93,8 +101,10 @@ func (im *ImprovedNoise) OffsetX() float64 { return im.xo }
 func (im *ImprovedNoise) OffsetY() float64 { return im.yo }
 func (im *ImprovedNoise) OffsetZ() float64 { return im.zo }
 
-// p mirrors ImprovedNoise.p(int): p[i & 255] & 255.
-func (im *ImprovedNoise) pp(i int) int { return im.p[i&255] & 255 }
+// p mirrors ImprovedNoise.p(int): p[i & 255] & 255. The stored value is a uint8
+// (already 0..255); widening to int and re-applying & 255 yields the identical
+// result the [256]int table returned.
+func (im *ImprovedNoise) pp(i int) int { return int(im.p[i&255]) & 255 }
 
 // Noise samples the noise at (x,y,z) with no y-fade (the 3-arg overload).
 func (im *ImprovedNoise) Noise(x, y, z float64) float64 {
